feat: operating on search query request realisation
