+++ src/const/const.ts
@@ -1,3 +1,5 @@
+// export const PREFIX = 'https://search.imdbot.workers.dev/?q=';
+
 export const INITIAL_STATE = [{
   name: '',
   isLogged: true
