+++ src/interfaces/films.interface.ts
@@ -1,0 +1,7 @@
+export interface IFilms {
+    id: number;
+    title: string;
+    isFavorite: boolean;
+    poster: string;
+    rating: number;
+}
