--- src/types/index.ts
@@ -1,7 +1,0 @@
-export interface IFilms {
-  id: number;
-  title: string;
-  isFavorite: boolean;
-  poster: string;
-  rating: number;
-}
